Catch throwables when instantiating interface
When documenting a laravel-json-api/laravel api I was getting a `Illuminate\Contracts\Container\BindingResolutionException` and the docs were failing to build. This PR catches those exceptions.

To give specifics: one of the arguments in the controller method is `LaravelJsonApi\Contracts\Routing\Route` (an interface). In the context of `knuckleswtf/scribe` that interface is not instantiating because `app()` doesn't have a binding to the interface.