Add deprecation date support
Before, the deprecation badge was just a boolean. Now it can be used both as a boolean
and a string. With a string, the badge is enhanced with said string to add some context.
This does not work with external resources since the OpenAPI specification accepts only
a boolean value.

Fix #1018